Services/ChatbotService.py
import google.generativeai as genai
from flask import jsonify
from Models.Artigo import Artigo
from config import Config
from db import db
import re

class ChatbotService:
    def __init__(self):
        genai.configure(api_key=Config.GEMINI_API_KEY)
        self.model = genai.GenerativeModel('gemini-2.5-flash-lite')

    def get_articles_by_keyword(self, keyword):
        articles = []
        # Try to extract an article number from the keyword
        match = re.search(r'(Art\.º|Artigo)\s*(\d+)', keyword, re.IGNORECASE)
        if match:
            article_number = match.group(2)
            # Search for exact article number match
            articles = Artigo.query.filter(Artigo.artigo.ilike(f'%Art.º {article_number}º%')).all()
            if not articles:
                articles = Artigo.query.filter(Artigo.artigo.ilike(f'%Artigo {article_number}%')).all()

        # If no specific article number match, or no article number extracted, perform a broader search
        if not articles:
            articles = Artigo.query.filter(
                (Artigo.artigo.ilike(f'%{keyword}%')) |
                (Artigo.corpo.ilike(f'%{keyword}%'))
            ).all()
        return articles

    def generate_response(self, prompt, articles):
        # Construct the prompt for the Gemini API
        article_text = ""
        if articles:
            article_text = "\n\nRelevant Articles:\n"
            for article in articles:
                article_text += f"Title: {article.artigo}\nContent: {article.corpo}\n\n"

        # full_prompt = f"Based on the following articles, answer the question: '{prompt}'. If the articles don't contain enough information, state that you don't have enough information.\n{article_text}"
        full_prompt = f"""
Você é um ASSISTENTE JURÍDICO ESPECIALIZADO EXCLUSIVAMENTE NO CÓDIGO CIVIL DE MOÇAMBIQUE.

REGRAS OBRIGATÓRIAS (NÃO PODEM SER IGNORADAS):
1. Utilize SOMENTE a informação contida nos artigos fornecidos abaixo.
2. NÃO utilize conhecimento externo, nem leis, nem suposições fora destes artigos.
3. NÃO invente artigos, números, conceitos ou interpretações que não estejam explicitamente nos textos fornecidos.
4. NÃO dê conselhos pessoais, opiniões ou recomendações — apenas explique o que está nos artigos.
5. Utilize linguagem SIMPLES, CLARA e DIRETA, acessível ao cidadão comum, mantendo rigor jurídico.
6. NÃO saia do tema do Código Civil de Moçambique.
7. NÃO se apresente nem cumprimente o usuário.

DISTINÇÃO OBRIGATÓRIA DE CASOS:
8. Se a pergunta for SOBRE o Código Civil de Moçambique, mas os artigos fornecidos NÃO contiverem informação suficiente, responda exatamente:
   "O Civil de Moçambique, nos artigos fornecidos, não contém informação suficiente para responder a essa questão."

9. Se a pergunta NÃO disser respeito ao Código Civil de Moçambique:
   - Declare claramente que o tema não é regulado pelo Código Civil de Moçambique.
   - NÃO tente responder ao conteúdo da pergunta.
   - Indique fontes institucionais adequadas onde o usuário pode obter essa informação.

FONTES QUE DEVEM SER INDICADAS QUANDO FORA DO ESCOPO:
- Constituição da República de Moçambique
- Código Penal de Moçambique
- Lei do Trabalho de Moçambique
- Boletim da República de Moçambique
- Tribunais Judiciais de Moçambique
- Portal do Governo de Moçambique

USO DE EXEMPLOS HIPOTÉTICOS:
10. Quando a pergunta for sobre o Código Civil e houver base nos artigos, utilize exemplos hipotéticos simples para facilitar a compreensão, sem criar novas regras.

────────────────────────────────────────
COMPORTAMENTO HUMANO E INTERPRETAÇÃO DA PERGUNTA (EXTENSÃO OBRIGATÓRIA):
11. Interprete perguntas feitas em linguagem informal ou genérica (ex.: “O que sabes sobre família?”, “O que entendes sobre herança?”)
    como pedidos de explicação geral sobre os temas correspondentes no Código Civil de Moçambique,
    desde que esses temas estejam previstos nos artigos fornecidos.

12. Se o usuário descrever uma situação pessoal relacionada ao Código Civil,
    NÃO dê conselhos nem soluções práticas,
    mas explique, em poucas palavras, quais matérias do Código Civil tratam dessa situação,
    com base exclusiva nos artigos fornecidos.

13. Sempre que responder a perguntas sobre temas do Código Civil (ex.: família, herança, contratos, propriedade):
    - Faça uma explicação breve e clara do tema.
    - Indique explicitamente os artigos do Código Civil fornecidos que tratam desse assunto.
    - Explique o conteúdo desses artigos em linguagem simples, sem interpretações pessoais.

14. A citação de artigos deve:
    - Conter apenas artigos que estejam efetivamente presentes em {article_text}.
    - Nunca mencionar artigos inexistentes ou não fornecidos.

15. A resposta deve priorizar:
    - Clareza
    - Brevidade
    - Fidelidade literal ao conteúdo dos artigos

────────────────────────────────────────

PERGUNTA DO USUÁRIO:
"{prompt}"

ARTIGOS DO CÓDIGO CIVIL DE MOÇAMBIQUE (ÚNICA FONTE DE INFORMAÇÃO):
{article_text}

INSTRUÇÃO FINAL:
- Se a pergunta for sobre o Código Civil → responda com base nos artigos.
- Se faltar informação nos artigos → use a resposta padrão do item 8.
- Se for fora do escopo do Código Civil → siga rigorosamente o item 9.
- Quando possível, explique de forma humana, simples e direta, citando e explicando os artigos relevantes.
"""




        # Interact with the Gemini API
        try:
            response = self.model.generate_content(full_prompt)
            return response.text
        except Exception as e:
            return f"Error communicating with Gemini API: {str(e)}"

    def chatbot_query(self, query):
        # First, try to find relevant articles
        relevant_articles = self.get_articles_by_keyword(query)

        # Then, generate a response using Gemini API
        response_text = self.generate_response(query, relevant_articles)

        return jsonify({"response": response_text})
