requirements.txt
alembic==1.9.1
apispec==6.0.2
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.1
flask-smorest==0.40.0
Flask-SQLAlchemy==3.0.2
greenlet>=3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==23.0
passlib==1.7.4
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.46
webargs==8.2.0
Werkzeug==2.2.2
google-generativeai
